Refactor UploadDispatcher to take Payloads
