+++ src/HolzShots.Core/Net/UploadUI.cs
@@ -10,55 +10,15 @@ using HolzShots.Drawing;
 
 namespace HolzShots.Net
 {
-    public interface IUploadPayload : IDisposable
-    {
-        string MimeType { get; }
-        string Extension { get; }
-        Stream GetStream();
-        string GetSuggestedFileName();
-    }
-
-    // TODO: refactor to use Bitmap
-    public record ImageUploadPayload : IUploadPayload
-    {
-        public string MimeType { get; init; }
-        public string Extension { get; init; }
-
-        private readonly Image _image;
-        private readonly ImageFormat _format;
-
-        public ImageUploadPayload(Image image, ImageFormat format)
-        {
-            _format = format ?? throw new ArgumentNullException(nameof(format));
-            _image = image.CloneGifBug(_format);
-
-            (Extension, MimeType) = _format.GetExtensionAndMimeType();
-            Debug.Assert(!string.IsNullOrWhiteSpace(MimeType));
-            Debug.Assert(!string.IsNullOrWhiteSpace(Extension));
-        }
-
-        public Stream GetStream() => _image.GetImageStream(_format);
-
-        public string GetSuggestedFileName()
-        {
-            Debug.Assert(Extension != null);
-            Debug.Assert(MimeType != null);
-
-            return ImageFormatInformation.DefaultUploadFileNameWithoutExtension + Extension;
-        }
-
-        public void Dispose() => _image.Dispose();
-    }
-
     public sealed class UploadUI : IDisposable
     {
-        private readonly ImageUploadPayload _payload;
+        private readonly IUploadPayload _payload;
         private readonly Uploader _uploader;
         private readonly ITransferProgressReporter? _progressReporter;
 
         private readonly SpeedCalculatorProgress _speedCalculator = new SpeedCalculatorProgress();
 
-        public UploadUI(ImageUploadPayload payload, Uploader uploader, ITransferProgressReporter? progressReporter)
+        public UploadUI(IUploadPayload payload, Uploader uploader, ITransferProgressReporter? progressReporter)
         {
             _payload = payload ?? throw new ArgumentNullException(nameof(payload));
             _uploader = uploader ?? throw new ArgumentNullException(nameof(uploader));
